@ARHAT : Imtegrated the image transfer code but not working

+++ src/model/ImageMetaData.java
@@ -8,7 +8,7 @@ public class ImageMetaData implements Serializable {
     private final int seq_no = -1;
     private int transmission_type = -1;
     private String file_name = "";
-    private int file_length = -1;
+    private long file_length = -1;
 
     public int getClient_id() {
         return client_id;
@@ -26,11 +26,11 @@ public class ImageMetaData implements Serializable {
         this.file_name = file_name;
     }
 
-    public int getFile_length() {
+    public long getFile_length() {
         return file_length;
     }
 
-    public void setFile_length(int file_length) {
+    public void setFile_length(long file_length) {
         this.file_length = file_length;
     }
